fix: Update esbuild-register package to use commonjs
