+++ packages/esbuild-register/src/hook.mts
