+++ packages/esbuild-register/src/index.cts
@@ -1,0 +1,3 @@
+import { register } from "./register.js";
+
+register();
