+++ packages/esbuild-register/src/index.mts
@@ -1,0 +1,3 @@
+import { register } from "./register.js";
+
+register();
